+++ admin_site/os2borgerpc_admin/initialize.py
@@ -20,7 +20,7 @@ def initialize():
     if os.path.exists(fixtures_dir) and os.path.isdir(fixtures_dir):
         initialize_sites()
         initialize_users()
-        initialize_distributions()
+        initialize_configuration_entries()
 
 
 def initialize_users():
@@ -49,14 +49,12 @@ def initialize_sites():
         call_command("loaddata", "sites.json", app_label="system")
 
 
-def initialize_distributions():
-    """Prime the system with some distributions to get started.
+def initialize_configuration_entries():
+    """Prime the system with some users to get started.
 
-    Data should be the output of "manage.py dumpdata system.Distribution"
-    and "manage.py dumpdata system.PackageList".
+    Data should be the output of
+    "manage.py dumpdata system.ConfigurationEntry".
+
     """
-    if os.path.exists(os.path.join(fixtures_dir, "distributions.json")):
-        call_command(
-            "loaddata", "packagelists.json", app_label="system"
-        )
-        call_command("loaddata", "distributions.json", app_label="system")
+    if os.path.exists(os.path.join(fixtures_dir, "configuration_entries.json")):
+        call_command("loaddata", "configuration_entries.json", app_label="system")