+++ admin_site/system/admin.py
@@ -10,15 +10,9 @@ from hashlib import md5
 from system.models import (
     Configuration,
     ConfigurationEntry,
-    PackageList,
-    Package,
     Site,
-    Distribution,
     PCGroup,
     PC,
-    CustomPackages,
-    PackageInstallInfo,
-    PackageStatus,
     ImageVersion,
     SecurityEvent,
     SecurityProblem,
@@ -30,32 +24,15 @@ from system.models import (
     AssociatedScript,
     AssociatedScriptParameter,
     ScriptTag,
+    Citizen,
 )
 
 ar = admin.site.register
 
 
-class PackageInstallInfoInline(admin.TabularInline):
-    model = PackageInstallInfo
-    extra = 3
-
-
-class PackageStatusInline(admin.TabularInline):
-    model = PackageStatus
-    extra = 3
-
-
 class ConfigurationEntryInline(admin.TabularInline):
     model = ConfigurationEntry
     extra = 3
-
-
-class PackageListAdmin(admin.ModelAdmin):
-    inlines = [PackageStatusInline]
-
-
-class CustomPackagesAdmin(admin.ModelAdmin):
-    inlines = [PackageInstallInfoInline]
 
 
 class SiteInlineForConfiguration(admin.TabularInline):
@@ -263,14 +240,15 @@ class AssociatedScriptAdmin(admin.ModelAdmin):
     search_fields = ("script__name",)
 
 
+class CitizenAdmin(admin.ModelAdmin):
+    list_display = ("citizen_id", "last_successful_login", "site")
+    search_fields = ("citizen_id",)
+
+
 ar(Configuration, ConfigurationAdmin)
-ar(PackageList)
-ar(CustomPackages, CustomPackagesAdmin)
 ar(Site, SiteAdmin)
-ar(Distribution)
 ar(PCGroup, PCGroupAdmin)
 ar(PC, PCAdmin)
-ar(Package)
 ar(ImageVersion, ImageVersionAdmin)
 # Job related stuff
 ar(Script, ScriptAdmin)
@@ -282,3 +260,4 @@ ar(AssociatedScript, AssociatedScriptAdmin)
 ar(AssociatedScriptParameter)
 ar(SecurityEvent, SecurityEventAdmin)
 ar(SecurityProblem, SecurityProblemAdmin)
+ar(Citizen, CitizenAdmin)